fix(materialize): gate the final operation's own overflow, and type the propagate zero
F1: the arithmetic BETWEEN the per-operand gates and the publish cast
was uncovered — a subtraction or division can overflow its OWN Spark
result type and yield NULL under ANSI-off with both operands healthy
(minuend 9.99e35, subtrahend -9.99e35, both DECIMAL(38,2): the
Subtract's DECIMAL(38,2) result cannot hold 1.998e36). Both per-operand
gates pass and the cast gate needs a pre-cast non-null, so the NULL
published as an empty window. A third clause now runs after the final
operation and before the operand drops: both operand value columns
non-null (and, under zero_denominator=null only, not denominator_is_zero
— the one policy that legitimately answers NULL there; zero answers with
a literal 0 and error already proved no zero denominator) yet the result
NULL raises OVERFLOW_VIOLATION naming the operation. The multi-slot
publish-cast comment now names all three checks, conditionally on which
were actually emitted.

M4: the propagate count's zero literal is typed (F.lit(0).cast('bigint')
— the COUNT's type, not the published one) per the file's _typed_literal
doctrine.

spark_semantics_gate.py gains the operation-level repro on the real
engine: the rendered DIFFERENCE node must refuse where each operand fits
and each sum is healthy but their difference does not fit — proven green
under both pyspark 3.5.1 and 4.2.0 via make l0-gate, with a fitting
control publishing first. Six calculation goldens regenerated.

+++ tests/featuregen/materialize/goldens/calculation_nodes/sum_decimal_propagate_nulls.py
@@ -36,7 +36,7 @@ def calculate_total_debit_amount_30d(
     # beside a count above zero — can only be Spark overflowing INSIDE the aggregation
     # (`CheckOverflowInSum` answers it with NULL before any cast). Dropped once the gate has read
     # it.
-    operand_count = F.when(any_null, F.lit(0)).otherwise(F.count(F.col('txn_amt')))
+    operand_count = F.when(any_null, F.lit(0).cast('bigint')).otherwise(F.count(F.col('txn_amt')))
 
     # The grain-level aggregate. `sum` is the expression's own declared aggregate, and the
     # grouping is the DECLARED grain — one row per landing key, which is what the spine reduction
